chore: support --yes on forge reset

+++ apps/forge-cli/src/cli/node/reset/reset.js
@@ -4,7 +4,7 @@ const inquirer = require('inquirer');
 const { config } = require('core/env');
 const { symbols } = require('core/ui');
 
-async function main() {
+async function main({ opts: { yes } }) {
   // Confirm stopped
   const { starterBinPath, forgeConfigPath } = config.get('cli');
   const { stdout: pid } = shell.exec(`FORGE_CONFIG=${forgeConfigPath} ${starterBinPath} pid`, {
@@ -20,17 +20,22 @@ async function main() {
   }
 
   // Confirm
-  const questions = [
-    {
-      type: 'confirm',
-      name: 'confirm',
-      default: false,
-      message: chalk.red(
-        'Reset chain state will erase all local data, including transactions/logs, are you sure to continue?'
-      ),
-    },
-  ];
-  const { confirm } = await inquirer.prompt(questions);
+  let confirm = yes;
+  if (!yes) {
+    const questions = [
+      {
+        type: 'confirm',
+        name: 'confirm',
+        default: false,
+        message: chalk.red(
+          'Reset chain state will erase all local data, including transactions/logs, are you sure to continue?'
+        ),
+      },
+    ];
+    const answers = await inquirer.prompt(questions);
+    confirm = answers.confirm;
+  }
+
   if (confirm) {
     shell.exec('rm -rf ~/.forge_release');
     shell.echo(`${symbols.info} rm -rf ~/.forge_release`);
